our_stuff/requirements.txt
ai2thor==5.0.0
ai2thor-colab==0.1.2
aws-requests-auth==0.4.3
blinker==1.8.2
botocore==1.35.53
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
decorator==4.4.2
Deprecated==1.2.14
Flask==3.0.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
moviepy==1.0.3
msgpack==1.1.0
numpy==2.1.2
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
prior==1.0.3
proglog==0.1.10
progressbar2==4.5.0
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.0
python-xlib==0.33
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.3.0
six==1.16.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.0
wrapt==1.16.0
